requirements.txt
blinker==1.9.0
cachetools==5.5.2
click==8.1.8
contourpy==1.3.2
cramjam==2.10.0
cubinlinker-cu11==0.3.0.post3
cuda-bindings==11.8.7
cuda-python==11.8.7
cudf-cu11==25.4.0
cupy-cuda11x==13.4.1
cycler==0.12.1
fastparquet==2024.11.0
fastrlock==0.8.3
filelock==3.13.1
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
fsspec==2025.3.0
greenlet==3.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
libcudf-cu11==25.4.0
libkvikio-cu11==25.4.0
librmm-cu11==25.4.0
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numba==0.61.2
numba-cuda==0.4.0
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.21.5
nvidia-nvcomp-cu11==4.2.0.11
nvidia-nvtx-cu11==11.8.86
nvitop==1.5.0
nvtx==0.2.11
packaging==25.0
pandas==2.2.3
pillow==11.0.0
psutil==7.0.0
ptxcompiler-cu11==0.8.1.post3
pyarrow==19.0.1
Pygments==2.19.1
pylibcudf-cu11==25.4.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rapids-logger==0.1.1
rich==14.0.0
rmm-cu11==25.4.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
SQLAlchemy==2.0.40
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.12.2
tzdata==2025.2
Werkzeug==3.1.3
